refactor: add test suite for SystemSettings, including model behavior and API permissions
